requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file='.\requirements\requirements.txt' '.\requirements\requirements.in'
#
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
build==1.2.2.post1
    # via pip-tools
click==8.1.7
    # via pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
django==5.1.2
    # via
    #   -r .\requirements\requirements.in
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==4.6.0
    # via -r .\requirements\requirements.in
django-rest-framework==0.1.0
    # via -r .\requirements\requirements.in
djangorestframework==3.15.2
    # via django-rest-framework
packaging==24.1
    # via build
pip-tools==7.4.1
    # via -r .\requirements\requirements.in
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
sqlparse==0.5.1
    # via django
tomli==2.0.2
    # via
    #   build
    #   pip-tools
typing-extensions==4.12.2
    # via asgiref
tzdata==2024.2
    # via django
wheel==0.44.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
